Add engine for schema
Sometimes, the schema needs to be generated dynamically. An example is
remote resources with paths dependent on environment variables.

So, this adds an engine/route/controller to do so, and changes the
install script to point to that endpoint. This means those who omit
vandal still get a schema hosted in `public` with
backwards-compatibility checks as normal, but vandal users will
reference a dynamically-generated schema.